requirements.txt
discord.py
spacy
requests
python-dotenv
pytest
pytest-asyncio
pytest-cov
requests-mock
python-dotenv
pytest-env
pytest-xdist